media: bcm2835-unicam: Use g_mbus_config to get the number of active lanes
Mainline kernels don't have a way of passing the number of active
lanes between the sensor and receiver drivers, they currently
rely on DT describing the one true configuration.
This doesn't work for the TC358743 HDMI to CSI-2 bridge (and
probably others) that need to configure the number of lanes to
roughly match the data rate of the incoming video.

Reintroduce the solution proposed (but objected to) upstream
as they still haven't decide how they want to solve the problem.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>