media: bcm2835-unicam: Check the sensor pixel format on G_FMT
With controls such as H&V flips on Bayer sensors they can
change the Bayer order, and hence the V4L2 pixel format.

The only path to reading back that value is via G_FMT, so
rather than returning the cached value of the pixel format,
check that the sensor format hasn't changed from the last
time, or update if it has. Attempt to match whether the
packed or unpacked format was previously in use with the new
format.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>