media: imx219: add PIXEL_RATE, VBLANK, and HBLANK controls
VBLANK is used to control frame rate.
When VBLANK value changes, recalculate the range for the exposure control
so that changing exposure wouldn't affect the frame rate.

PIXEL_RATE is the same in all the modes, so this is read-only control.

Line length is the same in all the modes, hence HBLANK depends on image
width only, and can only be changed by switching the mode. So HBLANK
is read-only control as well.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>