media: i2c: imx290: reorganize the code in preparation for new features
Reorganize and split sensor config values into more arrays in preparation
to extend the driver to handle:
  * 2 MIPI data lanes configuration in addition to the current 4 lanes one
  * RAW12 in addition to RAW10 pixel format
Also tweak the sensor configuration to:
  * make the 720p mode to use 60 fps (like the 1080p mode does)
  * set SHS1 register to 0x00001 (this gives the maximum integration time).
    W/o this change SHS1 register held the value of 0 which is prohibited
    according to the data sheet).
The good news is that with SHS1 being 1, the exposure in both 1080p, and
720p modes is almost the same now. Unfortunately this would no longer be
the case if we add a control to change SHS1 - changing SHS1 by the same
amount has different effect on exposure for 1080p and 720p.

Fix the MIPI link frequencies to match the datasheet (though they seem to
contradict the fps*HMAX*VMAX numbers).

Calculate pixel rate from the link frequency instead of using a constant:
according to the datasheet the current 1080p and 720p modes use different
link frequencies, not the same one.

Modify the link frequency check in imx290_probe() so that it passed if
all the link frequencies used by the driver are supported in the hardware
(present in the dtb).

Remove v4l2_fwnode_endpoint field from struct imx290 as it only needed in
the probe function: use the local variable for this purpose.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>